Cover more cases for Elixir queen attack exercise.
This adds more tests to cover the cases with queens placed on the edge
of the board or in the bottom right corner. After some thinking I did
about the cases I have extraced the closing character into a seperate
pattern matching function group. In the end I added one extra function
and was able to cover all the cases.